Move build_bank function to bank_handler file
